+++ src/lib/Menu.js
@@ -27,6 +27,7 @@ import { workspaceActions } from '../features/workspaces/actions';
 import { workspaceStore } from '../features/workspaces/index';
 import apiBase, { termsBase } from '../api/apiBase';
 import { openExternalUrl } from '../helpers/url-helpers';
+import globalMessages from '../i18n/globalMessages';
 
 const menuItems = defineMessages({
   edit: {
@@ -253,10 +254,6 @@ const menuItems = defineMessages({
   autohideMenuBar: {
     id: 'menu.app.autohideMenuBar',
     defaultMessage: '!!!Auto-hide menu bar',
-  },
-  quit: {
-    id: 'menu.app.quit',
-    defaultMessage: '!!!Quit',
   },
   addNewService: {
     id: 'menu.services.addNewService',
@@ -599,6 +596,23 @@ export default class FranzMenu {
     const tpl = _titleBarTemplateFactory(intl, this.stores.settings.app.locked);
     const { actions } = this;
 
+    // TODO: Extract this into a reusable component and remove the duplications
+    const quitApp = () => {
+      const yesButtonIndex = 0;
+      let selection = yesButtonIndex;
+      if (window.ferdi.stores.settings.app.confirmOnQuit) {
+        selection = dialog.showMessageBoxSync(app.mainWindow, {
+          type: 'question',
+          message: intl.formatMessage(globalMessages.quit),
+          detail: intl.formatMessage(globalMessages.quitConfirmation),
+          buttons: [intl.formatMessage(globalMessages.yes), intl.formatMessage(globalMessages.no)],
+        });
+      }
+      if (selection === yesButtonIndex) {
+        app.quit();
+      }
+    };
+
     if (!isMac) {
       tpl[1].submenu.push({
         label: intl.formatMessage(menuItems.autohideMenuBar),
@@ -801,11 +815,9 @@ export default class FranzMenu {
           type: 'separator',
         },
         {
-          label: intl.formatMessage(menuItems.quit),
-          role: 'quit',
-          click() {
-            app.quit();
-          },
+          label: intl.formatMessage(globalMessages.quit),
+          accelerator: `${cmdOrCtrlShortcutKey()}+Q`,
+          click: quitApp,
         },
       ],
     });
@@ -862,12 +874,9 @@ export default class FranzMenu {
           type: 'separator',
         },
         {
-          label: intl.formatMessage(menuItems.quit),
-          role: 'quit',
+          label: intl.formatMessage(globalMessages.quit),
           accelerator: `${cmdOrCtrlShortcutKey()}+Q`,
-          click() {
-            app.quit();
-          },
+          click: quitApp,
         },
       ];