+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix issue while importing exported `ferdi.data` file while switching from one laptop to another when it had workspaces (#1874) 💖 @vraravam
 - Fix for 'Support' link not working (404 error) (#1806) 💖 @kytwb
+- Fix zoom actions executed on another services (#1867) 💖 @sad270
 
 # [v5.6.1-nightly.51](https://github.com/getferdi/ferdi/compare/v5.6.1-nightly.50...v5.6.1-nightly.51) (2021-09-06)